refactor: React履歴画面の構成を整理
- App.tsx から表示コンポーネントを分割
- 履歴の状態管理を useHistories に切り出し
- 価格表示や検索キーワード整形を utils に分離
- 移行状況を更新
