+++ chrome/src/react/utils/keyword.ts
@@ -1,0 +1,6 @@
+export const normalizeKeywords = (keywords: string): string[] => {
+  return keywords
+    .replace(/\u3000/g, ' ')
+    .split(' ')
+    .filter(keyword => keyword !== '');
+};
